requirements.txt
Django==1.6.1
South==0.8.4
argparse==1.2.1
binaryornot==0.2.0
cookiecutter==0.7.0
distribute==0.6.24
django-environ==0.2.1
honcho==0.5.0
wsgiref==0.1.2